Test on_abort and on_reset for pandaseqtriggerpart
In order to prevent different PandA designs from interfering with the
abort and reset procedures, we explicitly ignore errors arising from the
lack of seqReset as an export.

For some reason, the test code produces a KeyError when it can't find
the export, but the equivalent on a real PandA resulted in an
AttributeError. This may be a mistake on my part, but it is safe to just
check for both types of error.