Fixes triggering of step scans with outer axes
After the first inner scan, the first point of subsequent inner scans
previously triggered immediately even if the outer axes was still
moving. This was because a check of whether the first point was static
didn't allow for scans with outer axes, in the case of step scans where
all points are static. (Their upper and lower bounds are equal to their
positions.)

To ensure that the sequence tables are ready to receive the first
signal, when the motors are in position, they are now enabled in the
PreRunHook. That is unless the scan doesn't contain any motors in which
case they are still enabled in the RunHook.